README.MD
# FMDE
Face Mask Detection Email

Face Mask Detection algorithm that sends email to desired personnel with datetime of when the mask is taken off.
Added additional algorithm to send email when mask is off. The code is still a little buggy, when mask is off it pauses streaming to save photo in the chosen folder and sents email, loops this until mask is put back on.
The bugs can easily be fixed and will be updated when I get the time to thread the functions together and put addition safe-guards.


Special thanks to balajisrinivas and his YouTube channel, if you want to get thorough explaination I recommend subscribing to his YouTube channel.